fix(app, app-shell, app-shell-odd): block initial HTTP request until successful MQTT subscription (#15094)
Closes EXEC-429

Currently, the desktop app/ODD attempt to address the "missed update problem" in the following way: while we subscribe to a topic, we simultaneously GET whatever equivalent HTTP resource we just subscribed to. However, there's definitely a world (albeit a very small one) in which we receive the HTTP response, a server update occurs, the server publishes, and then we successfully subscribe to a topic. In this world, we've missed the update event.

Solve for this by simply refetching right after the client subscribe ACKs. We still keep the initial fetch on mount to keep latency low.