fix(api): Improve attached_modules spec
This was getting a bit unwieldy; prior commits had made it a dict of
lists of strings, and now this needed it to be a dict of lists of tuples
of two strings, and all of this was terrible. Instead now it's a dict of
lists of a specific dataclass, much nicer.

While I was doing that I noticed that I didn't get any type errors, and
that's because ThreadManager's constructor was erasing the types of
parameters that you gave it, so I made it use a ParamSpec.
Unfortunately, you can't alter wrapped function named arguments using
type machinery right now, so I also had to change the way we specify
nonblocking thread manager construction to use a wrapper function that
sets an attr, but now that's type safe too.