refactor(app, api-client, react-api-client): unify analysis and run record for CommandText use
After a run is terminal, we access the run record rather than the most recently completed analysis
in the command list in the run preview (or more aptly, run postview). This PR refactors the data
sent to CommandText and ultimately many of its children and utilities to include only the necessary
data from the run record if the run is terminal, or the most recent analysis if the run has not
started or is in progress.