fix(app-shell-odd): allow running offboard (#15105)
It's useful to be able to run the odd shell on your laptop for testing
flows, and this code baked in a bunch of directory structure
assumptions. These fixes hopefully tolerate the changes by falling back
to not scanning for file changes anymore, allowing you to once again run
make -C app dev-odd.