feat(app): add desktop modal for incompatible modules
feat(app): IncompatibleModules on desktop

Display a modal on the desktop app if incompatible modules are attached
to the robot you're currently viewing. This modal allows interaction
with the navbar but otherwise is displayed at any time that the app is
viewing a specific robot (i.e. prepare for run, device details) and that
robot has incompatible modules attached.