refactor(app): split InterventionModal to molecule
The implementation of the intervention modal (i.e., manual move labware)
in the desktop app was sort of manually open-coded in the intervention
modal organism. We're going to need modals styled in this way elsewhere,
so split it out into an app molecule with some overridable styles.

Also, make its background wrapper position: absolute instead of
position: fixed, for the same reasons as #15166 - when this modal is
hung off of the modal portal rather than the top portal, it should allow
interaction with the navbar and the breadcrumbs. This should not affect
the modal's use in the actual InterventionRequired organism, since in
the organism the modal is hung off of the top portal.

Closes RSQ-6