feat(shared-data, api): add absorbance plate reader definition and module control (#15167)
<!--
Thanks for taking the time to open a pull request! Please make sure
you've read the "Opening Pull Requests" section of our Contributing
Guide:


https://github.com/Opentrons/opentrons/blob/edge/CONTRIBUTING.md#opening-pull-requests

To ensure your code is reviewed quickly and thoroughly, please fill out
the sections below to the best of your ability!
-->

# Overview
This PR adds support for the Absorbance Plate Reader in the hardware
controller API. We can now see the plate reader as an attached module on
the backend and use the AbsorbanceReaderDriver to communicate with the
plate reader like we would with the other modules.

# Caveat
The driver is not fully finished - there are some functions that still
need to be implemented and I expect the format of the response will
change depending on how we implement the engine commands.

We still need to add some changes on oe-core side to make sure this runs
on all robot - however i'm happy to show you the plate reader in action
at this time.