refactor(app): split InterventionModal to molecule (#15174)
The implementation of the intervention modal (i.e., manual move labware)
in the desktop app was sort of manually open-coded in the intervention
modal organism. We're going to need modals styled in this way elsewhere,
so split it out into an app molecule with some overridable styles.

Also, make its background wrapper position: absolute instead of
position: fixed, for the same reasons as #15166 - when this modal is
hung off of the modal portal rather than the top portal, it should allow
interaction with the navbar and the breadcrumbs. This should not affect
the modal's use in the actual InterventionRequired organism, since in
the organism the modal is hung off of the top portal.

Closes RSQ-6

## Testing
- [x] The new storybook component should render
- [x] Move labware and pause modals on desktop should still render as in
~figma~ the shipping app
- [x] Move labware and pause modals on desktop should continue to ~(a)
visually overlap and (b)~ inhibit interaction with the navbar and
breadcrumbs

the shipping app has these modals confined to the route component area,
while figma has them overlapping everything with a small padding to the
viewport boundary. this pr doesn't change that. the navbar and route
breadcrumbs are still non interactive.