feat(app): Add incompatible module notification modals (#15181)
We meant to add this for the launch of the Flex, but it got cut for
scope; in the time since, people connecting modules like
first-generation temperature modules to Flexes that don't support them
has been a pretty consistent thorn in the side of support. This PR
implements the modals designed by design for telling the user when
there's incompatible modules.

This PR is pretty big and might be best viewed by commit, since it has
both the backend and the frontend, and both the ODD and the desktop -
and a couple ancillary support PRs that I'll rebase out of here as they
get merged. Specifically,

## Overview

90dc748fbb - Python, adds data to module configs to say what kinds of
deck they're compatible with, and adds code to the api and robot server
to expose a bool about whether the pipette's compatible or not in the
HTTP API
788e022bac - JS, adds the above to the api clients
2e49f6160c - JS, adds a blocking modal to the ODD that pops up whenever
modules are attached per
[figma](https://www.figma.com/file/0kuPeCi1t2Auu2GPMnqRb2/Primary%3A-Flex?type=design&node-id=19961-155919&mode=design&t=pkbLPA9150bwQQNw-4)
c5021ca145 - JS, adds a modal that blocks interaction with a robot but
_not_ the navbar or breadcrumbs whenever modules are attached per
[figma](https://www.figma.com/design/Nx7ORMnfyJNP4FQYxN4e5x/Primary%3A-Desktop-App?node-id=2032-304206&t=iat6IH7YQD4pMijc-4)


## Notes and Review requests

90dc748fbb - I'm pretty sure these are the module compatibility rules,
but we'll need to make sure. This will get exposed via HTTP as soon as
the module is plugged in. Review requests:
- [ ] bool should be correct for plugged-in modules basically as soon as
they appear in the app

The above, and 788e022bac - This is set up to support polling. Let me
know if you think this should be notifications instead, will be another
couple commits though.

2e49f6160c - ODD modal. Is this according to design, is there something
I'm missing? Is the testing approach OK? You can test by editing
`robot-server/simulator/test-flex.env` to specify a
`ThermocyclerModuleV1` instead of a `ThermocyclerModuleV2` and run with
`make -C robot-server dev-flex`, as well as actually plugging stuff in
to a real robot.

c5021ca145 - Desktop modal, basically same questions and same testing
approach. Note that this is the first ish time we've added a modal that
doesn't inhibit interaction with navbar or breadcrumbs.