feat(api): set overlap versions from commands
PE commands that load or alter pipette configuration now allow
specifying a version of the pipette tip overlap dictionary to use. This
specification is on a "no version above <x>" basis, for future proofing.

It's allowed to not specify a version; in this case, the _highest_
available tip overlap version is used as a default.

This is different from similar implementations. It is done so that users
of protocol interfaces that do not require bug-for-bug compatibility as
the python protocol API does get the positioning changes for free -
those protocols do not have to change to get the new positioning.
Consistent behavior here is an opt-in. This will be true of JSON
protocols that use the protocol engine, v7 and v8.

The python protocol API, however, has to stay stable. For this reason,
- The hardware controller continues to export v0 of tip overlap
unconditionally in its pipette_dict
  - This guarantees same behavior for pre-engine protocol APIs
  - This guarantees same behavior for old robot server interfaces (i.e.
  OT-2 calibration)
  - This guarantees same behavior for old JSON protocol interfaces,
  which we don't actually care about
- The _engine_, on the other hand, actively loads the latest values if a
version isn't specified, which will happen when ingesting v7 and v8 PD
protocols
- The _engine core_ currently hard specifies v0, and that will change
based on the API level when we actually add a new version of the overlap data.

This is implemented at pipette configuration time so that it can stay
stateless - the pipette configuration data that is loaded in the engine
continues to present only one dictionary of tip overlap data, which is
resolved at the point of loading.