feat(api,robot-server,shared-data): Specify tip overlap by version (#15323)
We need to have multiple versions of the tip overlap data for at least
flex pipettes so that we can ship updated overlap data without
inadvertently changing the behavior of protocols on stable API versions.
This PR accomplishes that.


# Details
## separate versioning
The data in shared-data has a version, but that version is separate from
versions in other parts of the stack. This is on purpose, because the
version has to be written in shared-data, and read in protocol engine. I
think it's gross to write API versions in shared data, but we've done it
before and could do it again; the real sticking point was the engine.
The engine, in modern protocols, loads the tip overlap and does all that
data management itself; it would need to select the appropriate version
from the dict. We really don't want API versions there.

Therefore, there's a separate simple versioning scheme that's a little
more like the alternate liquid handling function scheme that we haven't
used in years, but with an actual string key for the version instead of
implicit versioning by positioning in an array.

There will instead be a specific map of API versions to overlap versions
living inside the source of the engine core for the python protocol API
(that's the only place it needs to be - for why, read on).

## engine state management and data lifecycle
There were basically two ways to implement this. The first way is to
save all the data - save the full versioned tip overlap dictionary, and
save which version we've selected. Since I listed this first, you can
guess I didn't do this. Instead, what I did was add a version to the two
engine commands that load pipette data - `load_pipette` and
`configure_for_volume` - and then have the functions those commands call
pull the specified version out of the full config, which means that we
only have to save the specific version in state and state doesn't have
to change. This also means that we don't have to worry about data
storage migration and compatibility.

One huge note here: **the engine argument parsing gives you the MOST
RECENT tip overlap if you don't specify one**. That's in bold because
it's not the way we've implemented similar things. It's necessary
because we want PD protocols to get the new overlap values without
having to change anything.

## python api integration
The python API, however, really wants to keep the old values (until
you're on a certain version - we'll implement that when we actually add
the new values). We do that in a couple ways:
- the old pipette dictionary value that has the tip overlap dict is
still there, using "v0". That means that protocol API versions that use
that dict, which is all those not using the engine core, get the old
values consistently for free
- for protocol api versions that do use the engine core, for now we hard
specify v0.

## json protocol integration

New JSON protocols, any that directly create engine commands rather than
being mapped through python protocol API values, will get the new
overlaps because they won't specify the parameter to the load pipette or
configure for volume engine commands. Older JSON protocols that do adapt
the python protocol API commands will get the python behavior.

# Testing
- This PR gets regression tested - it doesn't really contain any new
features, because it doesn't really contain any new data.
- OT-2 protocol behavior should be adequately covered by the g-code
testing
  - OT-2 calibration should probably be run real quick
- Flex behavior should be adequately tested by running LPC and a
protocol - this applies to all protocols and can be easily tested.
- Flex calibration and LPC doesn't use tips and therefore isn't affected


Closes EXEC-451