chore(app): add one col simple buttons component (#15366)
This is a subcomponent that can provide modal content for
InterventionModal. It lives inside the InterventionModal molecule (at
least for now). It's a radio button list where you can select one of the
provided buttons.

The argument passing style is a little weird but it's because there must
be at least two buttons, and this is a nice way to make it hard to do
anything else. They aren't child nodes because the whole point of this
component is to enforce a single button style.

Figma:

https://www.figma.com/design/8dMeu8MuPfXoORtOV6cACO/Feature%3A-Error-Recovery?node-id=2523-42677&t=yaBnNzMY2BoDzZZm-0

Storybook:

https://s3-us-west-2.amazonaws.com/opentrons-components/exec-490-one-col-simple-buttons/index.html?path=%2Fdocs%2Fdesign-tokens-borderradius--docs

Closes [EXEC-490](https://opentrons.atlassian.net/browse/EXEC-490)

[EXEC-490]:
https://opentrons.atlassian.net/browse/EXEC-490?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ