fix(app): ODD: Fix empty command text at start of run
CurrentRunningProtocolCommand is guaranteed to render only if robotSideAnalysis exists, but
CurrentRunningProtocolCommand assumes robotSideAnalysis could be undefined. Addtionally, based on
our use of .find() for obtaining the current command, typescript did not properly account for the
fact that the current command could possibly be null. While we do handle the null case, we do so by
rendering no command text. However, because we actually are guaranteed robotSideAnalyis, we can and
should render the first command, caching the idx. This is safe, because the only time lastRunCommand is null in
practice is when the run first begins, so there's never a scenario in which we accidentally render
the first run command except at the start of the run. But if lastRunCommand is ever null, we fallback to the cached idx.