fix(api): add runner load error handling to analyze cli (#15400)
Closes AUTH-486

# Overview

Fixes the issues caught in #15211 

Previously, RTP validation occurred as part of a simulating runner's run
process. This run process executes the protocol inside a separate task,
and any errors raised while running the task are wrapped and presented
in the 'Rune result" as an analysis error.

Now that RTP validation is happening as part of the runner load process,
we have to handle the error raised in the load process in a similar way
as well. This PR adds this missing error handling in `analyze` cli
function.

# Test Plan

- [ ] Upload a protocol with incorrect RTP definition and verify that
the in-app analysis fails with the correct error and no traceback in the
error pop-up window.
- [ ] Verify that the static data of the above protocol- metadata,
requirements and files are populated in the failed analysis
- [ ] Verify that any other, non-RTP-related errors raised during the
runner load process are handled just like they would have in the
previous stable software.

# Risk assessment

Medium. Updates a critical part of in-app analysis, but is a fix and is
isolated to error handling of `runner.load` only.