fix(protocol-designer): reset the touchTip checkbox when labware changes (#15407)
closes RESC-273

Background: A user selected a labware where touch tip is allowed and checked the touch tip box. Then the user changed the labware to one where touch tip is not allowed. The touch tip check box was disabled but not actually de-selected. This resulted in an unwanted erroneous `touchTip` command to be generated in step-generation.

The fix: when the aspirate, dispense, or mix labware changes, the touch tip checkbox field is reset. This fix is fine however if a protocol is already experiencing this error, it will not be automatically fixed upon importing their protocol.

The alternative fix: ensure that a touch tip command is only generated if the touch tip checkbox is selected AND there is a height is selected. That would mean changing the current behavior of PD.