feat(app, shared-data): check for new 'file-required' analysis result on protocol card and details
To accommodate file requirements in analysis failure, robot server will now introduce a 'result'
field on CLI analysis and update the existing 'result' field on robot analysis to provide a value of
'file-required'. This will be returned in the special case of analysis failure due to analysis
occurring before a CSV runtimeparameter value is input. Here, I update the these analyses'
respective interfaces accordingly. I also check for this status to render a "csv required" banner on
ProtocolCard and ProtocolDetails, and to prompt a warning message under protocol MiniCards on
ChooseProtocolSlideout.

Closes AUTH-448