feat(api, robot-server): return parameter-value-required result for analyses that reference un-set csv parameters (#15434)
# Overview

Closes AUTH-418 and addresses part of AUTH-419

This PR adds a new `result` type to analysis for the circumstance where
you reference a CSV parameter that is not set (since there are no local
defaults currently). In this scenario, the result will read
`parameter-value-required` instead of `ok` or `not-ok`. Since local
analysis did not previously have this field, it was added to local/cli
analysis as well.

In order to do this, a new error (and accordant error code) was added
that is raised when a csv parameter is accessed when it has not been
associated with a file. To accomplish this part, a basic CSV parameter
interface was added that currently stores a file (accessible via `.file`
with the parameter) and allows access to the rows (via `.rows`). Trying
to use either one of these without a file will raise a
`RuntimeParameterRequired` error.

This error is also re-exported to the `protocol_api` for public use in
the API as `RuntimeParameterRequiredError`. This can be caught and
handled, allowing a user to potentially provide their own in-protocol
alternate if they wish to have local analysis without providing a file
via a local dev-server.

# Test Plan

Tested that this protocol returns with a `parameter-value-required`
result in the analysis.

```
metadata = {
    'protocolName': 'CSV use parameter test',
}

requirements = {
    "robotType": "Flex",
    "apiLevel": "2.18"
}


def add_parameters(parameters):
    parameters.add_int(
        display_name="Sample count",
        variable_name="sample_count",
        default=2,
        minimum=1,
        maximum=6,
        description="How many samples to process."
    )
    parameters.add_str(
        display_name="Pipette Name",
        variable_name="pipette",
        choices=[
            {"display_name": "Single channel 50µL", "value": "flex_1channel_50"},
            {"display_name": "Single channel 1000µL", "value": "flex_1channel_1000"},
            {"display_name": "Eight Channel 50µL", "value": "flex_8channel_50"},
            {"display_name": "Eight Channel 1000µL", "value": "flex_8channel_1000"},
        ],
        default="flex_1channel_50",
        description="What pipette to use during the protocol.",
    )
    parameters.add_csv_file(
        display_name="CSV Data",
        variable_name="csv_data",
        description="CSV file containing labware and volume information."
    )



def run(context):
    SAMPLE_COUNT = context.params.sample_count
    PIPETTE_NAME = context.params.pipette

    trash_bin = context.load_trash_bin('A3')
    tip_rack = context.load_labware('opentrons_flex_96_tiprack_50ul', 'D2')

    pipette = context.load_instrument(PIPETTE_NAME, mount="left", tip_racks=[tip_rack])
    csv_file = context.params.csv_data.file
```

and that this correctly passes analysis

```
from opentrons import protocol_api

metadata = {
    'protocolName': 'CSV use parameter test w/ try except clause',
}

requirements = {
    "robotType": "Flex",
    "apiLevel": "2.18"
}


def add_parameters(parameters):
    parameters.add_csv_file(
        display_name="CSV Data",
        variable_name="csv_data",
        description="CSV file containing labware and volume information."
    )



def run(context):
    try:
    	csv_file = context.params.csv_data.file
    except protocol_api.RuntimeParameterRequiredError:
    	pass
```

# Changelog

- added new analysis `result` type of `parameter-value-required`
- added `result` field to local CLI analysis
- added a basic CSV interface to access file and rows. If there is no
file set, accessing those will raise a `RuntimeParameterRequired error.

# Review requests

Had to move some of the `code_in_error_tree` logic around to make it
importable. If that belong in a better place than `protocol_engine`, I
can move it, but left it there for now.

# Risk assessment

Medium-lowish, its a change to analysis (and particularly local analysis
with the new field), but its adding new information and shouldn't break
anything, even before official support of the `parameter-value-required`
field.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: jbleon95 <jbleon95@users.noreply.github.com>