refactor(app): update current step behavior (#15471)
Closes EXEC-572

Harmonizes the return value logic of currentStepNumber with totalStepCount.

Prior to 8.0, we’d fallback to deriving the current step count from the active run endpoint if we couldn’t get it from analysis (the run had diverged). Now, because we have fixit commands, we can’t do this, since that fallback step count will include fixit commands. This isn’t a functional change, just an "enforcing consistency change". We were already not using the run commands endpoint for “total steps” as a fallback. Realistically, we can always find the current step number from protocol analysis, since we filter out “not protocol” commands, so if currentStepNumber is ever null, there's an issue with a fetch, the protocol analysis itself, or most likely, the current command has diverged from analysis.