feat(components): StyledText styles for app/odd
StyledText is a component that is only really useful if you're goign to
do semantic styles. Semantic styles are only really useful if they mean
the same thing in all the places they're used. This isn't really true
right now, but we can at least have enumerated style tokens that you can
select from when creating one.

Now, which style token you use is separately selectable across ODD and
non-ODD flows. You can still fully or partially override the CSS if you
want, but the styles on desktop align with helix and the styles on ODD
align with what's going on in ODD.