feat(components): add helix product typography (#15459)
This PR implements the new helix product typography system. You can find
it here:
https://www.figma.com/design/1ot18My22DALIcjdLl5LJh/Helix-Design-System?node-id=3571-46578&t=B4pPbtL9mapki2Jg-4

The Product type system is distinct from at least the Web type system,
which will have different styles for readability in a different context.
It is currently only defined for places other than the ODD.

This ended up being a big PR because our usage of the `StyledText` atom
had to be changed somewhat.

The change was necessary because the helix product styles do not and are
not intended to align with the semantic styling presented by the
`StyledText` component - there is no direct alignment that "when you're
making a `<p>`, you use this style". Instead, the intent is that higher
level components will use the style by the name it has in helix, without
an explicit semantic link outside of the designs. That makes the way the
old styled text works kind of nonsense.

In addition, there are no helix product styles for the ODD, and there is
no current map and possibly no intention to have a future map between
the semantic styles of the ODD and Desktop. That means that having a
`StyledText` component that maps from a single semantic element to a set
of appropriate styles just doesn't really make sense.

So, `StyledText` is renamed to `LegacyStyledText` including at all
callsites. The new `StyledText` requires you to specify styles
separately for ODD and Desktop, though it gives you a handy type system
integrated hint towards which styles should be allowed. Well, not quite
required, more like you have no way to specify both in one prop.

 You can view it commit by commit:

-
https://github.com/Opentrons/opentrons/pull/15459/commits/07dae4b338d5b0966fc186fa27a944005a2e3fda
adds the actual data for the helix product typography styling, and
stories for viewing them as design components (and for viewing the old
typography styles). This is more or less the only commit there was
during the initial review, though it didn't then have the stories for
the old typography
-
https://github.com/Opentrons/opentrons/pull/15459/commits/a36f50178c93d95f9cf86610fe8313ae643afdaa
is a repo-wide find and replace to change `StyledText` to
`LegacyStyledText`
-
https://github.com/Opentrons/opentrons/pull/15459/commits/acec9340bcc955921c5007eb388ffec06ce6ac3c
implements the new `StyledText`.

Stories here:
https://s3-us-west-2.amazonaws.com/opentrons-components/EXEC-566-add-helix-product-typography/index.html?path=/docs/design-tokens-typography--docs

Closes EXEC-566